feat: remove local fallback to sqlite

+++ app/main.py
@@ -8,8 +8,7 @@ from starlette.requests import Request
 from starlette.responses import Response
 
 from app.common.config import (
-    DATABASE_RESET,
-    DB_URL,
+    DATABASE_MIGRATED,
     MQ_URL,
     PRODUCTION_MODE,
     Base,
@@ -24,12 +23,10 @@ from app.utils.cors import CorrectCORSMiddleware
 
 @asynccontextmanager
 async def lifespan(_: FastAPI) -> AsyncIterator[None]:
-    if not PRODUCTION_MODE:
+    if not PRODUCTION_MODE and not DATABASE_MIGRATED:
         async with engine.begin() as conn:
-            if DB_URL.endswith("app.db") or DATABASE_RESET:
-                await conn.run_sync(Base.metadata.drop_all)
-            if not DB_URL.startswith("postgresql") or DATABASE_RESET:
-                await conn.run_sync(Base.metadata.create_all)
+            await conn.run_sync(Base.metadata.drop_all)
+            await conn.run_sync(Base.metadata.create_all)
 
     loop: AbstractEventLoop = get_running_loop()
     connection = await connect_robust(MQ_URL, loop=loop)
